requirements.txt
wfastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
python-multipart==0.0.6
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
aiohttp==3.9.1 